requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
streamlit
python-dotenv
